refactor: mv validateRayClusterSpec to validation.go and mv related function
