+++ ray-operator/apis/ray/v1/constant.go
@@ -1,0 +1,3 @@
+package v1
+
+const RayFTEnabledAnnotationKey = "ray.io/ft-enabled"
