+++ ray-operator/apis/ray/v1/pod.go
@@ -1,0 +1,8 @@
+package v1
+
+import "strings"
+
+func IsGCSFaultToleranceEnabled(instance RayCluster) bool {
+	v, ok := instance.Annotations[RayFTEnabledAnnotationKey]
+	return (ok && strings.ToLower(v) == "true") || instance.Spec.GcsFaultToleranceOptions != nil
+}
